Text to Json
Feat: Implemented a text element to json converter, for saving the script.

Perf: Amended text element ID and all assignments to be an unsigned, 16 bit integer for maximal efficiency.

+++ Models/ScriptProject.cs
@@ -16,11 +16,11 @@ namespace Writers_Pal.Models
         public int RandomID() {
 
             Random rng = new Random();
-            int newID;
+            UInt16 newID;
 
             do {
 
-                newID = rng.Next(0, 65535);
+                newID = (ushort)rng.Next(0, 65535);
 
             } while (IsIDTaken(newID));
